backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.27.2
